% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angle}
\alias{angle}
\title{Torsion angles}
\format{
A data frame with 419 rows and 5 columns:
\describe{
\item{\code{chain}}{a character. The chain (either A or B) of the protein}
\item{\code{AA}}{a character. The three-letter amino acid name}
\item{\code{pos}}{an integer. Position in the amino acid sequence}
\item{\code{phi}}{numeric. The phi angle}
\item{\code{psi}}{numeric. The psi angle}
}
}
\source{
\url{https://warwick.ac.uk/fac/sci/moac/people/students/peter_cock/r/ramachandran}
}
\usage{
angle
}
\description{
The phi and psi torsion angles between peptide planes for the human protein
1HMP.
}
\keyword{datasets}
