% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{autoplot.tracer}
\alias{autoplot.tracer}
\alias{autoplot.trace}
\title{Plot results from a trace}
\usage{
\method{autoplot}{tracer}(object, y, ...)

\method{autoplot}{trace}(object, y, ...)
}
\arguments{
\item{object}{a trace or tracer object}

\item{y}{the name of the traced object to plot}

\item{...}{additional arguments passed to}
}
\value{

}
\description{
Plot results from a trace
}
