% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{greenland}
\alias{greenland}
\title{South west Greenland temperature data}
\format{
A data frame with 147 rows and 5 columns:
\describe{
\item{Year}{Year (numeric)}
\item{Month}{Month (encoded as a factor with levels 1 to 12)}
\item{Temp_Nuuk}{Montly mean temperature in Nuuk}
\item{Temp_Qaqortoq}{Montly mean temperature in Qaqortoq}
\item{Temp_diff}{Temperature difference, \code{Temp_Nuuk - Temp_Qaqortoq}}
}
}
\source{
\url{https://crudata.uea.ac.uk/cru/data/greenland}
}
\usage{
greenland
}
\description{
Monthly average temperatures in degree Celsius in Nuuk and Qaqortoq from 1873 to 2013.
}
\keyword{datasets}
