% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuuk}
\alias{nuuk}
\title{Nuuk temperature data}
\format{
A data frame with 147 rows and 6 columns:
\describe{
\item{Year}{Year (numeric)}
\item{Temperature}{Annual mean temperature}
\item{Median}{Annual median of montly mean temperatures}
\item{High}{Annual maximum of montly mean temperatures}
\item{Low}{Annual minimum of montly mean temperatures}
\item{Range}{High - Low}
}
}
\source{
\url{https://crudata.uea.ac.uk/cru/data/greenland}
}
\usage{
nuuk
}
\description{
Annual summaries of monthly mean temperatures in Nuuk from 1867 to 2013.
}
\keyword{datasets}
