% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminator.R
\name{print.terminator}
\alias{print.terminator}
\title{Print terminator information}
\usage{
\method{print}{terminator}(x, ...)
}
\arguments{
\item{x}{a terminator object}

\item{...}{other arguments (currently ignored)}
}
\value{
\code{print}
}
\description{
Print terminator information
}
