% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{[.tracer}
\alias{[.tracer}
\title{Subsetting tracer objects}
\usage{
\method{[}{tracer}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a tracer object.}

\item{i}{the indices of the trace information to extract.}

\item{j}{currently ignored.}

\item{...}{other arguments passed on to \code{get}.}

\item{drop}{currently ignored.}
}
\value{
a list
}
\description{
Subsetting tracer objects
}
